refactor(commands): sidebar editation
